@@ -4,17 +4,16 @@ build-backend = "flit_core.buildapi"
 
 [project]
 name = "oreum_core"
+dynamic = ["version", "description"]
 readme = "README.md"
 license = {file = "LICENSE.md"}
-authors = [
-    { name = "Oreum Industries", email = "info@oreum.io" }
-]
-requires-python = "==3.9.*"
+authors = [{ name = "Oreum Industries", email = "info@oreum.io" },]
+requires-python = "==3.10.*"
 classifiers = [
     "Development Status :: 4 - Beta",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
     "Intended Audience :: Science/Research",
     "Intended Audience :: Financial and Insurance Industry",
     "Topic :: Scientific/Engineering",
@@ -24,21 +23,24 @@ classifiers = [
     "Operating System :: Unix",
     "License :: OSI Approved :: Apache Software License",
 ]
-dynamic = ["version", "description"]
 dependencies = [
-    "arviz==0.14.*",   # NOTE v0.15 is incompatible due to removal of from_pymc3
     "csv2md>=1.1.2",
+    "fastparquet",   # not available in pandas v2.0 optional deps
     "ftfy>=5.4.1",   # NOTE MacOS has 6.1.1
     "graphviz<=0.18",   # NOTE capped https://stackoverflow.com/a/69997168/4473230
-    "matplotlib==3.5.2",
-    "matplotlib-inline==0.1.2",  # 0.1.6 causes massive plots in vscode https://github.com/ipython/ipython/issues/10383
-    "patsy>=0.5.0",  # TODO consider using formulae
-    "pymc3==3.11.4",  # TODO will soon upgrade to pymc v5
-    "seaborn==0.11.2",  # ignore the recent version with broken API
-    "theano-pymc==1.1.2",  # TODO will soon upgrade to pymc v5
-    "xarray<2023.2.0",  # upper pin to avoid issue https://github.com/Unidata/netcdf4-python/issues/1205, see also https://github.com/pydata/xarray/blob/main/doc/whats-new.rst#v2023020-feb-7-2023
-    "xlsxwriter>=3.0.3",
+    "matplotlib>=3.7.3",
+    "matplotlib-inline<0.1.5",  # 0.1.5+ causes massive plots in vscode https://github.com/ipython/ipython/issues/10383
+    "pandas[excel,parquet,plot]>=2.0.1",  # see NOTE
+    "patsy>=0.5.3",  # TODO consider using formulae instead
+    "pymc==5.6.1",  # >=5.8.2",  # brings along lots of deps
+    "seaborn==0.12.2",  # 0.13 brings many breaking changes to avoid right now
 ]
+
+# NOTE ON DEPS
+# + pymc: deps bring many packages incl.: arviz (xarray), numpy, pandas, pytensor, scipy, tables
+# + pandas: added minimal optional deps. importing projects might also consider
+#   including pandas[performance,postgresql]>=2.0.0
+
 
 [project.optional-dependencies]
 dev = [
@@ -47,8 +49,9 @@ dev = [
     "flake8",
     "hypothesis",
     "interrogate",
+    "ipython",
     "isort",
-    "nose",
+    "pipdeptree",
     "pip-licenses",
     "pre-commit",
     "pytest",
@@ -58,12 +61,8 @@ dev = [
 "Homepage" = "https://oreum.io"
 "Source Code" = "https://github.com/oreum-industries/oreum_core"
 
-[metadata]
-description-file = 'README.md'
-license_files = 'LICENSE.md'
-
-[bdist_wheel]
-universal=1
+# [bdist_wheel]
+# universal=1
 
 [tool.bandit]
 skips = [
