@@ -15,7 +15,7 @@
 """Core tools for use on projects by Oreum Industries"""
 import logging
 
-__version__ = "0.4.19"
+__version__ = "0.5.0"
 
 # logger goes to null handler by default
 # packages that import oreum_core can override this and direct elsewhere
