@@ -13,39 +13,32 @@
 # limitations under the License.
 
 # model/
-"""Various classes & functions for modelling, primarily using pymc3"""
-from oreum_core.model.base import BasePYMC3Model
+"""Various classes & functions for modelling, primarily using pymc"""
+from oreum_core.model.base import BasePYMCModel
 from oreum_core.model.calc import (
     calc_2_sample_delta_prop,
     calc_bayesian_r2,
     calc_binary_performance_measures,
-    calc_dist_fns_over_x,
-    calc_dist_fns_over_x_manual_only,
+    calc_f_beta,
     calc_mse,
     calc_ppc_coverage,
     calc_r2,
     calc_rmse,
+    compute_log_likelihood_for_potential,
     log_jcd,
     numpy_invlogit,
 )
 from oreum_core.model.describe import (
     describe_dist,
     extract_yobs_yhat,
-    get_posterior_summary,
+    get_summary,
     model_desc,
 )
 from oreum_core.model.distributions import (
-    InverseWeibull,
-    InverseWeibullNumpy,
-    Kumaraswamy,
-    Lognormal,
-    LognormalNumpy,
-    Normal,
-    NormalNumpy,
-    ZeroInflatedInverseWeibull,
-    ZeroInflatedLogNormal,
-    boundzero_numpy,
-    boundzero_theano,
+    lognormal_icdf,
+    mv_dist,
+    normal_icdf,
+    sanity_check_lognorm,
 )
 from oreum_core.model.model_io import ModelIO
 from oreum_core.model.plot import (
@@ -53,8 +46,9 @@ from oreum_core.model.plot import (
     forestplot_multiple,
     forestplot_single,
     pairplot_corr,
-    plot_dist_fns_over_x,
-    plot_dist_fns_over_x_manual_only,
-    plot_ppc_loopit,
+    plot_compare,
+    plot_energy,
+    plot_loo_pit,
+    plot_ppc,
     plot_trace,
 )
